+++ tests/test_01_validaciones.py
@@ -1,6 +1,6 @@
 from importlib.machinery import SourceFileLoader
 
-mod = SourceFileLoader("m", "src/01_validaciones.py").load_module()
+mod = SourceFileLoader("m", "src/validaciones.py").load_module()
 
 def test_es_texto_vacio():
     assert mod.es_texto_vacio("") is True
